requirements.txt
frida-tools>=14.0.0
pycryptodome>=3.20.0
zstandard>=0.22.0
